pyproject.toml
[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
force-exclude = """
src/pages/(.*).py |
config/globalparameter.py
"""

[tool.poetry]
name = "gf-qa"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "~3.10"
selenium = "^4.1.3"
rich = "^11.2.0"
pygsheets = "^2.0.5"
PyYAML = "^6.0"
pdfplumber = "^0.6.0"
behave = "^1.2.6"
behave-plain-color-formatter = "^0.0.2"
behave-html-formatter = "^0.9.8"
allure-behave = "^2.9.45"
allure-python-commons = "^2.9.45"
imutils = "^0.5.4"
opencv-python = "^4.5.5"
pandas = "^1.4.1"
scikit-image = "^0.19.2"
requests = "^2.27.1"
matplotlib = "^3.5.1"
slackclient = "^2.9.3"
nose = "^1.3.7"
NoseXML = "^0.19"
python-dateutil = "^2.8.2"
pyparsing = "^3.0.7"
beautifulsoup4 = "^4.10.0"
pywin32 = {version = "^303", platform = "win32"}
macApp = {version = "^1.1", platform = "darwin"}
openpyxl = "^3.0.9"
polling2 = "^0.5.0"
scipy = "1.8.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.17.0"
isort = "^5.10.1"
autoflake = "^1.4"
gitlint = "^0.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
